add refresh token api

+++ src/main/java/al/webgis/webgis/model/UserDto.java
@@ -1,4 +1,4 @@
 package al.webgis.webgis.model;
 
-public record UserDto(Integer id, String username) {
+public record UserDto(Long id, String username) {
 }
